requirements.txt
streamlit
tensorflow-cpu
Pillow
numpy
requests
